added methods of loading in external materials
