+++ neutronics_material_maker/__init__.py
@@ -1,1 +1,4 @@
-from .material_maker_classes import Material, MultiMaterial
+from .material import Material
+from .material import AddMaterialFromDir
+from .material import AddMaterialFromFile
+from .mutlimaterial import MultiMaterial
